+++ sanitize/src/decorators/index.ts
@@ -1,10 +1,13 @@
 
 import { UnSignedInt, SignedInt, RangedInt } from './integer';
 import { Bool } from './boolean';
+import { Uuid } from './uuid';
+import { NotNull } from './notnull';
 
 export default {
     UnSignedInt, 
     SignedInt, 
     RangedInt,
-    Bool
+    Bool,
+    Uuid,
 }